Update OpenSSL to 1.0.0k to resolve CVE-2013-0169
Without this patch the Ruby interpreter produced by this build system is
statically linked against OpenSSL 1.0.0j.  This is a problem because
this version of OpenSSL is vulnerable to remote attackers conducting
distinguishing attacks and plaintext-recovery attacks via statistical
analysis of timing data for crafted packets, aka the "Lucky Thirteen"
issue.

This is a problem for Puppet Labs because we use the Ruby provided by
this build system to distribute Puppet and Puppet Enterprise which both
rely on SSL and TLS for network security.

This patch addresses the problem by incrementing the version of OpenSSL
to 1.0.0k, which has addressed and resolved this vulnerability.